requirements.txt
asttokens==2.4.1
autopep8==2.0.4
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
dash==2.15.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.1
decorator==5.1.1
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
Flask==3.0.2
fonttools==4.49.0
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
ipykernel==6.29.2
ipython==8.18.1
ipywidgets==8.1.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
logging==0.4.9.6
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
numpy==1.26.3
openpyxl==3.1.2
packaging==23.2
pandas==2.2.0
parso==0.8.3
patsy==0.5.6
pillow==10.2.0
platformdirs==4.2.0
plotly==5.19.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pycodestyle==2.11.1
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
pyzmq==25.1.2
requests==2.31.0
retrying==1.3.4
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
stack-data==0.6.3
statsmodels==0.14.1
tenacity==8.2.3
threadpoolctl==3.3.0
tomli==2.0.1
tornado==6.4
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
wcwidth==0.2.13
Werkzeug==3.0.1
widgetsnbextension==4.0.10
yadisk==2.1.0
zipp==3.17.0
